[jsk_recognition_utils] Separate plane.h from geo_util.h
